Rename Deck and Panel classes
