+++ firefox_puppeteer/firefox_puppeteer/ui/update_wizard/wizard.py
@@ -5,7 +5,7 @@
 from marionette_driver import By, Wait
 
 from firefox_puppeteer.ui_base_lib import UIBaseLib
-from firefox_puppeteer.ui.deck import BasePanel as Panel
+from firefox_puppeteer.ui.deck import Panel
 
 
 class Wizard(UIBaseLib):
